Create jsonrpc2clientwallet.php
The addition of this file allows the wallet to use rpc credentials. Starting in version 0.10.0.1, monero-wallet-rpc requires a rpc username and password by default. If the user does not specify a username and password with the --rpc-login argument, a random rpc username and password is written to the same directory as monero-wallet-rpc with the name of monero-wallet-rpc.[walletPortNumber].login  If the user specifies a username and password with the --rpc-login argument, such as --rpc-login user:pass, that username and password will be used and written to monero-wallet-rpc.[walletPortNumber].login